# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Campus Resource Management and Booking System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rm/resources:
    get:
      operationId: rmListResources
      tags: [rm]
      summary: List resources managed by me
      responses:
        "200":
          description: Resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourceDetail"
    post:
      operationId: rmCreateResource
      tags: [rm]
      summary: Create a new resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceBody"
      responses:
        "201":
          description: Created resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceDetail"
  /rm/resources/{id}:
    patch:
      operationId: rmUpdateResource
      tags: [rm]
      summary: Update resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceBody"
      responses:
        "200":
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceDetail"
  /rm/resources/{id}/status:
    patch:
      operationId: rmUpdateResourceStatus
      tags: [rm]
      summary: Update resource status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResourceStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /rm/resources/{id}/unavailability:
    post:
      operationId: rmAddUnavailability
      tags: [rm]
      summary: Add maintenance window
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUnavailabilityBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceUnavailability"
  /rm/resources/{id}/unavailability/{uid}:
    delete:
      operationId: rmDeleteUnavailability
      tags: [rm]
      summary: Delete maintenance window
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: uid
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /rm/approvals/{id}/decide:
    post:
      operationId: rmDecideApproval
      tags: [rm]
      summary: Approve or reject step-1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionBody"
      responses:
        "200":
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    ResourceDetail:
      type: object
      properties:
        resource_id:
          type: integer
        resource_name:
          type: string
        capacity:
          type: integer
        location:
          type: ["string", "null"]
        status:
          type: string
        features:
          type: object
        category_id:
          type: integer
        category_name:
          type: string
        approval_steps:
          type: integer
        manager_id:
          type: ["integer", "null"]
        manager_name:
          type: ["string", "null"]
        department_id:
          type: ["integer", "null"]
        department_name:
          type: ["string", "null"]
        unavailability:
          type: array
          items:
            $ref: "#/components/schemas/ResourceUnavailability"
      required: [resource_id, resource_name, capacity, status, features, category_id, category_name, approval_steps]
...
    CreateResourceBody:
      type: object
      properties:
        resource_name:
          type: string
        capacity:
          type: integer
        location:
          type: ["string", "null"]
        status:
          type: string
        features:
          type: object
        category_id:
          type: integer
        department_id:
          type: ["integer", "null"]
      required: [resource_name, capacity, category_id]
    UpdateResourceStatusBody:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ResourceUnavailability:
      type: object
      properties:
        unavail_id:
          type: integer
        resource_id:
          type: integer
        day_of_week:
          type: ["integer", "null"]
        start_time:
          type: string
        end_time:
          type: string
        label:
          type: ["string", "null"]
      required: [unavail_id, resource_id, start_time, end_time]
    CreateUnavailabilityBody:
      type: object
      properties:
        day_of_week:
          type: ["integer", "null"]
        start_time:
          type: string
        end_time:
          type: string
        label:
          type: ["string", "null"]
      required: [start_time, end_time]
...
    ApprovalDecisionBody:
      type: object
      properties:
        decision:
          type: string
        remarks:
          type: ["string", "null"]
      required: [decision]
...