# lib/api-zod/src/generated/types/createResourceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Resource Management and Booking System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateResourceBodyFeatures } from "./createResourceBodyFeatures";

export interface CreateResourceBody {
  resource_name: string;
  capacity: number;
  /** @nullable */
  location?: string | null;
  status?: string;
  features?: CreateResourceBodyFeatures;
  category_id: number;
  /** @nullable */
  department_id?: number | null;
}