# lib/api-zod/src/generated/types/updateResourceStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Resource Management and Booking System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateResourceStatusBody {
  status: string;
}